+++ src/avatar/avatar.vue
@@ -1,29 +1,21 @@
 <template>
-  <div :class="avatarClass" :style="customSize">
-    <div :class="`${name}__inner`">
-      <t-image
-        v-if="image && !hideOnLoadFailed"
-        :style="customSize"
-        v-bind="customImageProps"
-        @load="handleImgLoadCompleted"
-        @error="handleImgLoadError"
-      />
-      <div v-else-if="iconContent !== undefined" :class="`${name}__icon`">
-        <t-node :content="iconContent"></t-node>
-      </div>
-      <span v-else>
-        <t-node :content="avatarContent"></t-node>
-      </span>
-    </div>
-    <div v-if="badgeProps" :class="`${name}__badge`">
-      <t-badge
-        :count="badgeProps.count"
-        :max-count="badgeProps.maxCount"
-        :dot="badgeProps.dot"
-        :content="badgeProps.content"
-        :size="badgeProps.size"
-        :offset="badgeProps.offset"
-      >
+  <div :class="`${name}__wrapper`" :style="customSize">
+    <div :class="`${name}__badge`">
+      <t-badge v-bind="customBadgeProps">
+        <div :class="avatarClass">
+          <t-image
+            v-if="image && !hideOnLoadFailed"
+            :style="customSize"
+            v-bind="customImageProps"
+            @error="handleImgLoadError"
+          />
+          <div v-else-if="iconContent !== undefined" :class="`${name}__icon`">
+            <t-node :content="iconContent"></t-node>
+          </div>
+          <span v-else>
+            <t-node :content="avatarContent"></t-node>
+          </span>
+        </div>
       </t-badge>
     </div>
   </div>
@@ -51,15 +43,19 @@ export default defineComponent({
     const emitEvent = useEmitEvent(props, context.emit);
     const internalInstance = getCurrentInstance();
     const avatarGroupProps = inject('avatarGroup', {}) as TdAvatarGroupProps;
+    const hasAvatarGroupProps = Object.keys(avatarGroupProps).length > 0;
     const iconContent = computed(() => renderTNode(internalInstance, 'icon'));
     const avatarContent = computed(() => renderContent(internalInstance, 'default', 'content'));
     const sizeValue = ref(props.size || (avatarGroupProps && avatarGroupProps.size));
+    const sizeReValue = ref((avatarGroupProps && avatarGroupProps.size) || props.size);
+    const sizeClass = `${sizeReValue.value.indexOf('px') > -1 ? 'medium' : sizeReValue.value}`;
     const avatarClass = computed(() => [
       `${name}`,
-      sizeValue.value ? CLASSNAMES.SIZE[sizeValue.value] : '',
+      `${name}--${sizeClass}`,
       {
         [`${name}--${props.shape}`]: props.shape,
       },
+      hasAvatarGroupProps ? `${name}--border ${name}--border-${sizeClass}` : '',
     ]);
 
     const isCustomSize = computed(() => sizeValue.value && !CLASSNAMES.SIZE[sizeValue.value]);
@@ -71,12 +67,14 @@ export default defineComponent({
           }
         : {};
     });
-    const handleImgLoadCompleted = (e: any) => {
-      emitEvent('load', e);
-    };
+
     const handleImgLoadError = (e: any) => {
       emitEvent('error', e);
     };
+
+    const customBadgeProps = computed(() => ({
+      ...props.badgeProps,
+    }));
 
     const baseImageProps = {
       src: props.image,
@@ -94,9 +92,9 @@ export default defineComponent({
       avatarContent,
       avatarClass,
       customSize,
-      handleImgLoadCompleted,
       handleImgLoadError,
       customImageProps,
+      customBadgeProps,
     };
   },
 });