+++ src/collapse/demos/placement.vue
@@ -1,0 +1,7 @@
+<template>
+  <t-collapse>
+    <t-collapse-panel value="0" placement="top" header="折叠面板标题">
+      此处可自定义内容此处可自定义内容此处可自定义内容此处可自定义内容此处可自定义内容此处可自定义内容此处可自定义内容此处可自定义内容
+    </t-collapse-panel>
+  </t-collapse>
+</template>