+++ src/components.ts
@@ -24,7 +24,7 @@ export { default as Message } from './message';
 export { default as DateTimePicker } from './date-time-picker';
 export { default as Toast } from './toast';
 export { default as Search } from './search';
-export { Steps, Step } from './steps';
+export { Steps, StepItem } from './steps';
 export { Indexes, IndexesCell } from './indexes';
 export { Picker, Cascade } from './picker';
 export { Tag, CheckTag } from './tag';