+++ src/progress/style/index.js
@@ -1,1 +1,1 @@
-import '../../_common/style/mobile/components/progress/_index.less';
+import '../../_common/style/mobile/components/progress/v2/_index.less';