+++ src/steps/demos/horizontal.vue
@@ -1,61 +1,58 @@
 <template>
-  <div class="step-demo">
-    <t-steps v-model="current" :readonly="true">
-      <t-step title="当前步骤"></t-step>
-      <t-step title="未完成步骤"></t-step>
-    </t-steps>
-    <t-steps v-model="current" :readonly="true">
-      <t-step title="当前步骤" content="辅助信息文字最多两行"></t-step>
-      <t-step title="未完成步骤" content="辅助信息文字最多两行"></t-step>
-    </t-steps>
-    <t-steps v-model="current" :readonly="true">
-      <t-step title="错误步骤" content="辅助信息文字最多两行" status="error"></t-step>
-      <t-step title="未完成步骤" content="辅助信息文字最多两行"></t-step>
-    </t-steps>
-  </div>
-  <div class="step-demo">
-    <t-steps v-model="current1" :readonly="true">
-      <t-step title="已完成步骤"></t-step>
-      <t-step title="当前步骤"></t-step>
-      <t-step title="未完成步骤"></t-step>
-    </t-steps>
-    <t-steps v-model="current1" :readonly="true">
-      <t-step title="已完成步骤" content="辅助信息文字最多两行"></t-step>
-      <t-step title="当前步骤" content="辅助信息文字最多两行"></t-step>
-      <t-step title="未完成步骤" content="辅助信息文字最多两行"></t-step>
-    </t-steps>
-    <t-steps v-model="current1" :readonly="true">
-      <t-step title="已完成步骤" content="辅助信息文字最多两行"></t-step>
-      <t-step title="错误步骤" content="辅助信息文字最多两行" status="error"></t-step>
-      <t-step title="未完成步骤" content="辅助信息文字最多两行"></t-step>
-    </t-steps>
-  </div>
-  <div class="step-demo">
-    <t-steps v-model="current2" :readonly="true">
-      <t-step title="已完成步骤"></t-step>
-      <t-step title="当前步骤"></t-step>
-      <t-step title="未完成步骤"></t-step>
-      <t-step title="未完成步骤"></t-step>
-    </t-steps>
-    <t-steps v-model="current2" :readonly="true">
-      <t-step title="已完成步骤" content="辅助信息文字最多两行"></t-step>
-      <t-step title="当前步骤" content="辅助信息文字最多两行"></t-step>
-      <t-step title="未完成步骤" content="辅助信息文字最多两行"></t-step>
-      <t-step title="未完成步骤" content="辅助信息文字最多两行"></t-step>
-    </t-steps>
-    <t-steps v-model="current2" :readonly="true">
-      <t-step title="已完成步骤" content="辅助信息文字最多两行"></t-step>
-      <t-step title="错误步骤" content="辅助信息文字最多两行" status="error"></t-step>
-      <t-step title="未完成步骤" content="辅助信息文字最多两行"></t-step>
-      <t-step title="未完成步骤" content="辅助信息文字最多两行"></t-step>
-    </t-steps>
+  <div>
+    <div class="summary">水平带序号步骤条</div>
+    <div class="steps-horizontal-demo-block">
+      <t-steps :current="options.first" @change="onFirstChange">
+        <t-step-item v-for="(_, index) in count" :key="index" :title="getTitle('first', index)" content="辅助信息" />
+      </t-steps>
+    </div>
+    <div class="summary">水平带图标步骤条</div>
+    <div class="steps-horizontal-demo-block">
+      <t-steps :current="options.second" @change="onSecondChange">
+        <t-step-item v-for="(_, index) in count" :key="index" :title="getTitle('second', index)" content="辅助信息">
+          <template #icon>
+            <CartIcon size="20px" />
+          </template>
+        </t-step-item>
+      </t-steps>
+    </div>
+    <div class="summary">水平简略步骤条</div>
+    <div class="steps-horizontal-demo-block">
+      <t-steps :current="options.third" theme="dot" @change="onThirdChange">
+        <t-step-item v-for="(_, index) in count" :key="index" :title="getTitle('third', index)" content="辅助信息" />
+      </t-steps>
+    </div>
   </div>
 </template>
+<script lang="ts" setup>
+import { reactive, ref } from 'vue';
+import { CartIcon } from 'tdesign-icons-vue-next';
 
-<script setup lang="ts">
-import { ref } from 'vue';
+const options = reactive({
+  first: 1,
+  second: 1,
+  third: 1,
+});
+const count = ref(4);
+const onFirstChange = (current: number) => {
+  options.first = current;
+};
+const onSecondChange = (current: number) => {
+  options.second = current;
+};
+const onThirdChange = (current: number) => {
+  options.third = current;
+};
 
-const current = ref(0);
-const current1 = ref(1);
-const current2 = ref(1);
+const getTitle = (type: 'first' | 'second' | 'third', index: number) => {
+  if (index === options[type]) {
+    return '当前步骤';
+  }
+  if (index < options[type]) {
+    return '已完成';
+  }
+  if (index > options[type]) {
+    return '未完成';
+  }
+};
 </script>